Kafka feeds added, Gauge module, Highstock updated, Dashboard mode, Parameters can be changed at runtime

+++ src/java/com/unifina/feed/MessageSource.java
@@ -4,5 +4,5 @@ package com.unifina.feed;
 
 public interface MessageSource {
 	public void setRecipient(MessageRecipient recipient);
-	public void setExpectedCounter(int expected);
+	public void setExpectedCounter(long expected);
 }
